models/inference/XGBoost-SM/requirements.txt
xgboost
pandas
numpy
huggingface_hub
psycopg2-binary
python-dotenv
